Enhance operation and team models with additional properties
- Updated `OperationDetails` to include `AdministratorIds`, `CreatedAt`, and `UpdatedAt` for improved operation management.
- Modified `Team` and `TeamRecord` classes to replace `ProfitShareRule` with `OperatorProfitShareRuleRecord`, enhancing clarity in profit sharing logic.
- Adjusted mappings and repository implementations to reflect the new structure and ensure consistency across the application.
- Improved the dashboard API types and added functionality for assigning and unassigning administrators to operations, enhancing user management capabilities.

+++ Nexus/Operations/Application/Services/TeamService.cs
@@ -91,7 +91,7 @@ public sealed class TeamService : ITeamService
             TeamLeaderId: null,
             OperatorIds: Array.Empty<string>(),
             StrawManIds: Array.Empty<string>(),
-            GatewaySelectionStrategy: (int)GatewaySelectionStrategy.PerStrawman,
+            GatewaySelectionStrategy: GatewaySelectionStrategy.PerStrawman,
             GatewayCredentialsIds: Array.Empty<string>(),
             GatewayCredentialsGroupIds: Array.Empty<string>(),
             OperatorProfitShareRules: Array.Empty<OperatorProfitShareRuleRecord>(),
